api/requirements.txt
fastapi
tensorflow == 2.15.0
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.15.0
matplotlib
numpy